Reworked Enums for Bible and Apocrypha (among other stuff)
- Now just one EnumBible and EnumApocrypha, rather than having separate
Aliases and Chapter/Verse count enums
- These enums now have JUnit tests, used to ensure that any changes to
these enums are caught in JUnit testing (e.g. if someone accidently
deletes a chapter or two from Psalms, or moves stuff around)
- The constants in the JUnit folder should not be used as API classes
(if there's a way for them to be). They are meant to be unchanging stuff
to be used for ensuring the actual constants (in the enums) are
consistent

- Version string changed to be 2.0-Alpha1-SNAPSHOT, to be more
representative of what's being worked towards (1.0 = Bukkit version
ultimately).

- Stuff changed for new enum changes